font + tailwind + vueanimejs (completly worked)

+++ database/migrations/2019_06_11_042118_create_praktikans_table.php
@@ -18,7 +18,7 @@ class CreatePraktikansTable extends Migration
             $table->string('nama', 50);
             $table->string('nim', 20)->unique();
             $table->string('password');
-            $table->integer('kelas_id');
+            $table->unsignedBigInteger('kelas_id');
             $table->text('alamat');
             $table->integer('nomor_telepon');
             $table->string('email')->unique();
